refactor(archive): move projection builder out of lib
