+++ devtools/build_topology_projection.py
@@ -107,8 +107,8 @@ LIB_PREFIX_TO_SUBPACKAGE = {
     "filter_": "archive/filter/",
     "filters": "archive/filter/",
     "phase_": "archive/phase/",
-    "projection_": "lib/projection/",
-    "projections": "lib/projection/",
+    "projection_": "archive/projection/",
+    "projections": "archive/projection/",
     "provider_": "archive/provider/",
     "attachment_": "archive/attachment/",
 }
@@ -212,7 +212,7 @@ TARGET_TO_OWNER = [
     ("polylogue/lib/conversation/", "lib-conversation"),
     ("polylogue/lib/semantic/", "lib-semantic"),
     ("polylogue/archive/phase/", "archive-phase"),
-    ("polylogue/lib/projection/", "lib-projection"),
+    ("polylogue/archive/projection/", "archive-projection"),
     ("polylogue/archive/provider/", "archive-provider"),
     ("polylogue/archive/attachment/", "archive-attachment"),
     ("polylogue/storage/repository/", "storage-repository"),
