+++ tests/unit/core/test_conversation_semantics.py
@@ -13,10 +13,10 @@ from datetime import datetime
 
 import pytest
 
+from polylogue.archive.projection.projections import ConversationProjection
 from polylogue.lib.message.messages import MessageCollection
 from polylogue.lib.models import Attachment, Conversation, DialoguePair, Message
 from polylogue.lib.pricing import harmonize_session_cost
-from polylogue.lib.projection.projections import ConversationProjection
 from polylogue.lib.roles import Role
 from tests.infra.assertions import assert_contains_all, assert_not_contains_any
 from tests.infra.builders import make_conv, make_msg
